requirements.txt
Flask==1.1.2
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.1.0
tensorflow==2.8.0
tensorflow_gpu==2.8.0
